Bump MongoDB PHP driver to v1.15.0 / Install PHP8.2 MongoDB extension (#1844)
* mongodb.sh: bump PHP driver to 1.14.1 version

* Bump MongoDB PHP driver to v1.15.0

* Install PHP8.2 MongoDB extension

* mongodb.sh: fix copy'n'paste issue

+++ scripts/features/mongodb.sh
@@ -41,7 +41,7 @@ sudo systemctl enable mongod
 sudo systemctl start mongod
 
 sudo rm -rf /tmp/mongo-php-driver /usr/src/mongo-php-driver
-git clone -c advice.detachedHead=false -q -b '1.14.2' --single-branch https://github.com/mongodb/mongo-php-driver.git /tmp/mongo-php-driver
+git clone -c advice.detachedHead=false -q -b '1.15.0' --single-branch https://github.com/mongodb/mongo-php-driver.git /tmp/mongo-php-driver
 sudo mv /tmp/mongo-php-driver /usr/src/mongo-php-driver
 cd /usr/src/mongo-php-driver
 git submodule -q update --init
@@ -140,14 +140,27 @@ then
     sudo service php8.0-fpm restart
 fi
 
-phpize8.1
-./configure --with-php-config=/usr/bin/php-config8.1 > /dev/null
+if [ -f /home/$WSL_USER_NAME/.homestead-features/php81 ]
+then
+    phpize8.1
+    ./configure --with-php-config=/usr/bin/php-config8.1 > /dev/null
+    make clean > /dev/null
+    make >/dev/null 2>&1
+    sudo make install
+    sudo bash -c "echo 'extension=mongodb.so' > /etc/php/8.1/mods-available/mongo.ini"
+    sudo ln -s /etc/php/8.1/mods-available/mongo.ini /etc/php/8.1/cli/conf.d/20-mongo.ini
+    sudo ln -s /etc/php/8.1/mods-available/mongo.ini /etc/php/8.1/fpm/conf.d/20-mongo.ini
+    sudo service php8.1-fpm restart
+fi
+
+phpize8.2
+./configure --with-php-config=/usr/bin/php-config8.2 > /dev/null
 make clean > /dev/null
 make >/dev/null 2>&1
 sudo make install
-sudo bash -c "echo 'extension=mongodb.so' > /etc/php/8.1/mods-available/mongo.ini"
-sudo ln -s /etc/php/8.1/mods-available/mongo.ini /etc/php/8.1/cli/conf.d/20-mongo.ini
-sudo ln -s /etc/php/8.1/mods-available/mongo.ini /etc/php/8.1/fpm/conf.d/20-mongo.ini
-sudo service php8.1-fpm restart
+sudo bash -c "echo 'extension=mongodb.so' > /etc/php/8.2/mods-available/mongo.ini"
+sudo ln -s /etc/php/8.2/mods-available/mongo.ini /etc/php/8.2/cli/conf.d/20-mongo.ini
+sudo ln -s /etc/php/8.2/mods-available/mongo.ini /etc/php/8.2/fpm/conf.d/20-mongo.ini
+sudo service php8.2-fpm restart
 
 mongosh admin --eval "db.createUser({user:'homestead',pwd:'secret',roles:['root']})"
